moving header-editor component to connected components
