+++ packages/connected-components/src/index.tsx
@@ -3,5 +3,6 @@ export {
   PureNewNotebookNavigation
 } from "./new-notebook-navigation";
 
+export { default as HeaderEditor } from "./header-editor";
 export { default as NotebookMenu } from "./notebook-menu";
 export { default as ModalController } from "./modal-controller";
