+++ packages/presentational-components/src/index.ts
@@ -1,6 +1,5 @@
 import { Cell } from "./components/cell";
 import { Cells } from "./components/cells";
-import { HeaderEditor } from "./components/header-editor";
 import { Input } from "./components/input";
 import { Outputs } from "./components/outputs";
 import { Pagers } from "./components/pagers";
@@ -9,14 +8,4 @@ import { Source } from "./components/source";
 
 export * from "./styles";
 
-export {
-  Input,
-  Outputs,
-  Pagers,
-  Prompt,
-  PromptBuffer,
-  Source,
-  Cell,
-  Cells,
-  HeaderEditor
-};
+export { Input, Outputs, Pagers, Prompt, PromptBuffer, Source, Cell, Cells };
